Some refactoring to better handle frozen states
